HMP-368: includes alt text in thumbnail link when image missing
