+++ extras/bampfa/app/controllers/catalog_controller.rb
@@ -13,9 +13,13 @@ class CatalogController < ApplicationController
     config.advanced_search[:query_parser] ||= 'edismax'
     config.advanced_search[:form_solr_parameters] ||= {}
 
+    config.index.thumbnail_presenter = ThumbnailPresenter
     config.view.gallery(document_component: Blacklight::Gallery::DocumentComponent)
     config.view.masonry(document_component: Blacklight::Gallery::DocumentComponent)
-    config.view.slideshow(document_component: Blacklight::Gallery::SlideshowComponent)
+    config.view.slideshow(
+      document_component: Blacklight::Gallery::SlideshowComponent,
+      preview_component: Gallery::SlideshowPreviewComponent
+    )
     config.show.tile_source_field = :content_metadata_image_iiif_info_ssm
     config.show.partials.insert(1, :openseadragon)
