+++ extras/cinefiles/app/controllers/catalog_controller.rb
@@ -17,6 +17,7 @@ class CatalogController < ApplicationController
     # config.view.gallery(document_component: Blacklight::Gallery::DocumentComponent)
     # config.view.masonry(document_component: Blacklight::Gallery::DocumentComponent)
     # config.view.slideshow(document_component: Blacklight::Gallery::SlideshowComponent)
+    config.index.thumbnail_presenter = ThumbnailPresenter
     config.show.tile_source_field = :content_metadata_image_iiif_info_ssm
     config.show.partials.insert(1, :openseadragon)
