+++ extras/pahma/app/assets/stylesheets/extras.scss
@@ -40,6 +40,7 @@
 }
 .section-heading {
   font-size: 1.25rem;
+  margin-top: 0.9rem;
 }
 .site-footer {
   color: white;
